+++ hutool-cron/src/main/java/com/xiaoleilu/hutool/cron/CronUtil.java
@@ -1,3 +1,4 @@
+
 package com.xiaoleilu.hutool.cron;
 
 import java.util.Map.Entry;
@@ -6,7 +7,9 @@ import com.xiaoleilu.hutool.convert.Convert;
 import com.xiaoleilu.hutool.cron.task.Task;
 import com.xiaoleilu.hutool.exceptions.UtilException;
 import com.xiaoleilu.hutool.setting.Setting;
+import com.xiaoleilu.hutool.setting.SettingRuntimeException;
 import com.xiaoleilu.hutool.util.ClassUtil;
+import com.xiaoleilu.hutool.util.CollectionUtil;
 
 /**
  * 定时任务工具类
@@ -37,7 +40,20 @@ public final class CronUtil {
 	 * @param cronSettingPath 定时任务配置文件路径（相对绝对都可）
 	 */
 	public static void setCronSetting(String cronSettingPath) {
-		crontabSetting = new Setting(cronSettingPath, Setting.DEFAULT_CHARSET, false);
+		try {
+			crontabSetting = new Setting(cronSettingPath, Setting.DEFAULT_CHARSET, false);
+		} catch (SettingRuntimeException e) {
+			//ignore setting file parse error
+		}
+	}
+	
+	/**
+	 * 是否使用精确到秒的定时，默认不使用
+	 * @param isSecondMode <code>true</code>支持，<code>false</code>不支持
+	 * @return this
+	 */
+	public static void setUseSecond(boolean isUseSecond) {
+		scheduler.setUseSecond(isUseSecond);
 	}
 	
 	/**
@@ -65,16 +81,17 @@ public final class CronUtil {
 	 * @param cronSetting 定时任务设置文件
 	 */
 	public static void schedule(Setting cronSetting) {
+		if(CollectionUtil.isEmpty(cronSetting)){
+			return;
+		}
 		for (Entry<Object, Object> entry : cronSetting.entrySet()) {
 			final String jobClass = Convert.toStr(entry.getKey());
 			final String pattern = Convert.toStr(entry.getValue());
 			try {
 				final Runnable job = ClassUtil.newInstance(jobClass);
 				schedule(pattern, job);
-//				log.info("Schedule [{} {}] added.", pattern, jobClass);
 			} catch (Exception e) {
-				e.printStackTrace();
-//				log.error(e, "Schedule [%s %s] add error!", pattern, jobClass);
+				throw new CronException(e, "Schedule [{}] [{}] error!", pattern, jobClass);
 			}
 		}
 	}