+++ hutool-log/src/main/java/com/xiaoleilu/hutool/log/LogFactory.java
@@ -9,42 +9,52 @@ import com.xiaoleilu.hutool.log.dialect.slf4j.Slf4jLogFactory;
 
 /**
  * 日志工厂类
+ * 
+ * @see Slf4jLogFactory
+ * @see Log4jLogFactory
+ * @see Log4j2LogFactory
+ * @see ApacheCommonsLogFactory
+ * @see JdkLogFactory
+ * @see ConsoleLogFactory
+ * 
  * @author Looly
  *
  */
 public abstract class LogFactory {
-	
+
 	private String logFramworkName;
-	
+
 	public LogFactory(String logFramworkName) {
 		this.logFramworkName = logFramworkName;
 	}
-	
+
 	/**
 	 * 获得日志对象
+	 * 
 	 * @param name 日志对象名
 	 * @return 日志对象
 	 */
 	public abstract Log getLog(String name);
-	
+
 	/**
 	 * 获得日志对象
+	 * 
 	 * @param clazz 日志对应类
 	 * @return 日志对象
 	 */
 	public abstract Log getLog(Class<?> clazz);
-	
-	//------------------------------------------------------------------------- Static start
+
+	// ------------------------------------------------------------------------- Static start
 	private static volatile LogFactory currentLogFactory;
 	private static final Object lock = new Object();
 
 	/**
 	 * @return 当前使用的日志工厂
 	 */
-	public static LogFactory getCurrentLogFactory(){
-		if(null == currentLogFactory){
+	public static LogFactory getCurrentLogFactory() {
+		if (null == currentLogFactory) {
 			synchronized (lock) {
-				if(null == currentLogFactory){
+				if (null == currentLogFactory) {
 					currentLogFactory = detectLogFactory();
 				}
 			}
@@ -54,57 +64,95 @@ public abstract class LogFactory {
 	
 	/**
 	 * 自定义日志实现
-	 * @param logFactory 日志工厂类
+	 * 
+	 * @see Slf4jLogFactory
+	 * @see Log4jLogFactory
+	 * @see Log4j2LogFactory
+	 * @see ApacheCommonsLogFactory
+	 * @see JdkLogFactory
+	 * @see ConsoleLogFactory
+	 * 
+	 * @param logFactoryClass 日志工厂类
 	 * @return 自定义的日志工厂类
 	 */
-	public static LogFactory setCurrentLogFactory(LogFactory logFactory){
+	public static LogFactory setCurrentLogFactory(Class<? extends LogFactory> logFactoryClass) {
+		try {
+			return setCurrentLogFactory(logFactoryClass.newInstance());
+		} catch (Exception e) {
+			throw new IllegalArgumentException("Can not instance LogFactory class!", e);
+		}
+	}
+
+	/**
+	 * 自定义日志实现
+	 * 
+	 * @see Slf4jLogFactory
+	 * @see Log4jLogFactory
+	 * @see Log4j2LogFactory
+	 * @see ApacheCommonsLogFactory
+	 * @see JdkLogFactory
+	 * @see ConsoleLogFactory
+	 * 
+	 * @param logFactory 日志工厂类对象
+	 * @return 自定义的日志工厂类
+	 */
+	public static LogFactory setCurrentLogFactory(LogFactory logFactory) {
 		logFactory.getLog(LogFactory.class).debug("Custom Use [{}] Logger.", logFactory.logFramworkName);
 		currentLogFactory = logFactory;
 		return currentLogFactory;
 	}
-	
+
 	/**
 	 * 获得日志对象
+	 * 
 	 * @param name 日志对象名
 	 * @return 日志对象
 	 */
-	public static Log get(String name){
+	public static Log get(String name) {
 		return getCurrentLogFactory().getLog(name);
 	}
-	
+
 	/**
 	 * 获得日志对象
+	 * 
 	 * @param clazz 日志对应类
 	 * @return 日志对象
 	 */
-	public static Log get(Class<?> clazz){
+	public static Log get(Class<?> clazz) {
 		return getCurrentLogFactory().getLog(clazz);
 	}
-	
+
 	/**
 	 * @return 获得调用者的日志
 	 */
 	public static Log get() {
 		return get(new Exception().getStackTrace()[1].getClassName());
 	}
-	
+
 	/**
 	 * @return 获得调用者的调用者的日志（用于内部辗转调用）
 	 */
 	protected static Log indirectGet() {
 		return get(new Exception().getStackTrace()[2].getClassName());
 	}
-	
+
 	/**
 	 * 决定日志实现
+	 * 
+	 * @see Slf4jLogFactory
+	 * @see Log4jLogFactory
+	 * @see Log4j2LogFactory
+	 * @see ApacheCommonsLogFactory
+	 * @see JdkLogFactory
+	 * @see ConsoleLogFactory
 	 * @return 日志实现类
 	 */
-	private static LogFactory detectLogFactory(){
+	private static LogFactory detectLogFactory() {
 		LogFactory logFactory;
-		try{
+		try {
 			logFactory = new Slf4jLogFactory(true);
 			logFactory.getLog(LogFactory.class).debug("Use [{}] Logger As Default.", logFactory.logFramworkName);
-		}catch(Throwable e){
+		} catch (Throwable e) {
 			try {
 				logFactory = new Log4jLogFactory();
 				logFactory.getLog(LogFactory.class).debug("Use [{}] Logger As Default.", logFactory.logFramworkName);
@@ -128,8 +176,8 @@ public abstract class LogFactory {
 				}
 			}
 		}
-		
+
 		return logFactory;
 	}
-	//------------------------------------------------------------------------- Static end
+	// ------------------------------------------------------------------------- Static end
 }